requirements.txt
anthropic
python-dotenv
fastapi[standard]
langchain_community
langchain_anthropic